chore: Refactor into Tokens struct
