+++ tests/tokens_test.rs
@@ -1,11 +1,11 @@
-use momoa::{tokenize, TokenKind, Location, Tokens};
+use momoa::{tokenize_json, tokenize_jsonc, TokenKind, Location};
 use test_case::test_case;
 
 #[test_case("false", TokenKind::Boolean ; "tokenize_false")]
 #[test_case("true", TokenKind::Boolean ; "tokenize_true")]
 #[test_case("null", TokenKind::Null ; "tokenize_null")]
 fn should_tokenize_keywords(code: &str, kind: TokenKind) {
-    let result = tokenize(code).unwrap();
+    let result = tokenize_json(code).unwrap();
     assert_eq!(result[0].kind, kind);
     assert_eq!(result[0].loc.start, Location {
         line: 1,
@@ -30,7 +30,7 @@ fn should_tokenize_keywords(code: &str, kind: TokenKind) {
 #[test_case("1" ; "tokenize_single_digit")]
 #[test_case("-1345.98324978324780943" ; "tokenize_negative_long_float")]
 fn should_tokenize_numbers(code: &str) {
-    let result = tokenize(code).unwrap();
+    let result = tokenize_json(code).unwrap();
     assert_eq!(result[0].kind, TokenKind::Number);
     assert_eq!(result[0].loc.start, Location {
         line: 1,
@@ -48,56 +48,56 @@ fn should_tokenize_numbers(code: &str) {
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_end_of_input_reading_minus() {
-    tokenize("-").unwrap();
+    tokenize_json("-").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_end_of_input_reading_float() {
-    tokenize("5.").unwrap();
+    tokenize_json("5.").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected character '1' found.")]
 fn should_panic_unexpected_start_of_number() {
-    tokenize("01").unwrap();
+    tokenize_json("01").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected character 'e' found.")]
 fn should_panic_unexpected_e() {
-    tokenize("-e").unwrap();
+    tokenize_json("-e").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected character '.' found.")]
 fn should_panic_unexpected_start_with_dot() {
-    tokenize(".1").unwrap();
+    tokenize_json(".1").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_end_after_e() {
-    tokenize("25e").unwrap();
+    tokenize_json("25e").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_plus_after_e() {
-    tokenize("3E+").unwrap();
+    tokenize_json("3E+").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_minus_after_e() {
-    tokenize("33e-").unwrap();
+    tokenize_json("33e-").unwrap();
 }
 
 #[test_case("\"hello\"" ; "simple")]
 #[test_case("\"hello\\u32AF\"" ; "unicode_escape")]
 #[test_case("\"\\n\\r\\t\\f\\\"\\b\"" ; "normal_escapes")]
 fn should_tokenize_strings(code: &str) {
-    let result = tokenize(code).unwrap();
+    let result = tokenize_json(code).unwrap();
     assert_eq!(result[0].kind, TokenKind::String);
     assert_eq!(result[0].loc.start, Location {
         line: 1,
@@ -114,37 +114,37 @@ fn should_tokenize_strings(code: &str) {
 #[test]
 #[should_panic(expected="Unexpected character 'X' found.")]
 fn should_panic_unexpected_unicode_escape_character() {
-    tokenize("\"hello\\u32AX\"").unwrap();
+    tokenize_json("\"hello\\u32AX\"").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected character '\"' found.")]
 fn should_panic_premature_unicode_escape_end() {
-    tokenize("\"hello\\u32A\"").unwrap();
+    tokenize_json("\"hello\\u32A\"").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unicode_escape_end_of_input() {
-    tokenize("\"hello\\u32A").unwrap();
+    tokenize_json("\"hello\\u32A").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected end of input found.")]
 fn should_panic_unexpected_end_of_string() {
-    tokenize("\"hello").unwrap();
+    tokenize_json("\"hello").unwrap();
 }
 
 #[test]
 #[should_panic(expected="Unexpected character 'x' found.")]
 fn should_panic_invalid_escape() {
-    tokenize("\"\\x\"").unwrap();
+    tokenize_json("\"\\x\"").unwrap();
 }
 
 #[test]
 fn should_tokenize_line_comment_without_eol() {
     let code = "// foo";
-    let result = tokenize(code).unwrap();
+    let result = tokenize_jsonc(code).unwrap();
     assert_eq!(result[0].kind, TokenKind::LineComment);
     assert_eq!(result[0].loc.start, Location {
         line: 1,
@@ -161,7 +161,7 @@ fn should_tokenize_line_comment_without_eol() {
 #[test]
 fn should_tokenize_line_comment_with_eol() {
     let code = "// foo\n";
-    let result = tokenize(code).unwrap();
+    let result = tokenize_jsonc(code).unwrap();
     assert_eq!(result[0].kind, TokenKind::LineComment);
     assert_eq!(result[0].loc.start, Location {
         line: 1,
@@ -173,32 +173,4 @@ fn should_tokenize_line_comment_with_eol() {
         column: code.len() + 1,
         offset: code.len()
     });
-}
-
-#[test]
-fn tokens_should_tokenize_code() {
-    let code = "true";
-    let mut tokens = Tokens::new(code);
-    let token = tokens.next().unwrap().unwrap();
-
-    assert_eq!(token.kind, TokenKind::Boolean);
-    assert_eq!(token.loc.start, Location {
-        line: 1,
-        column: 1,
-        offset: 0
-    });
-    assert_eq!(token.loc.end, Location {
-        line: 1,
-        column: code.len() + 1,
-        offset: code.len()
-    });
-
-}
-
-#[test]
-fn tokens_should_not_tokenize_empty_string() {
-    let mut tokens = Tokens::new("");
-    let result = tokens.next().unwrap();
-
-    assert!(result.is_none());
 }
